requirements.txt
flask==3.0.2
gunicorn==22.0.0
flask-cors
requests
python-dotenv